+++ news.md
@@ -1,6 +1,6 @@
 # skilljaR 0.1.0.9000
 
-* dev version
+* Created `get_course_paths()` which will retrieve the list of paths and then the items (courses) associated with each path.
 
 # skilljaR 0.1.0